requirements.txt
streamlit
tensorflow==2.12.0
opencv-python-headless
pillow
numpy
pandas
reportlab